+++ src/Foundation.Domain.Repositories/IReadOnlyGenericRepository.cs
@@ -56,14 +56,14 @@
         /// Counts all entities.
         /// </summary>
         /// <param name="cancellationToken">The cancellation token.</param>
-        Task<long> CountAsync(CancellationToken cancellationToken = default);
+        Task<int> CountAsync(CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Counts all entities.
         /// </summary>
         /// <param name="specification">The specification.</param>
         /// <param name="cancellationToken">The cancellation token.</param>
-        Task<long> CountAsync(
+        Task<int> CountAsync(
             ISpecification<TEntity> specification,
             CancellationToken cancellationToken = default);
 
@@ -72,7 +72,7 @@
         /// </summary>
         /// <param name="specifications">The specifications.</param>
         /// <param name="cancellationToken">The cancellation token.</param>
-        Task<long> CountAsync(
+        Task<int> CountAsync(
             IEnumerable<ISpecification<TEntity>> specifications,
             CancellationToken cancellationToken = default);
     }
